backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prism"
version = "1.0.0"
description = "The Prism — Local-first AI behavioral manifold tooling suite by BuildMaxxing"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "BuildMaxxing", email = "dev@buildmaxxing.com" }]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Validation & settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    # HuggingFace (config-only; no weights downloaded in Phase 1)
    "huggingface-hub>=0.27.0",
    # HTTP client (for license checks, HF metadata)
    "httpx>=0.28.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Rich output (shared with CLI in Phase 2)
    "rich>=13.9.0",
]

[project.optional-dependencies]
# Heavy ML stack — optional so the API server starts without GPU deps
ml = [
    "torch>=2.1.0",
    "transformers>=4.48.0",
    "peft>=0.14.0",
]
# Dev / test
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[project.scripts]
prism = "prism.cli.main:app"   # wired in Phase 2

[tool.setuptools.packages.find]
where = ["."]
include = ["prism*"]

# ---------------------------------------------------------------------------
# Ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["prism", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["prism"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "INFO"
